Factor out lifecycle script execution
