+++ tests/specs/npm/lifecycle_scripts/node_gyp_not_found.out
@@ -6,3 +6,4 @@ Initialize @denotest/node-addon-implicit-node-gyp@1.0.0
 warning: node-gyp was used in a script, but was not listed as a dependency. Either add it as a dependency or install it globally (e.g. `npm install -g node-gyp`)
 [WILDCARD]
 error: script 'install' in '@denotest/node-addon-implicit-node-gyp@1.0.0' failed with exit code 1
+error: failed to run scripts for packages: @denotest/node-addon-implicit-node-gyp@1.0.0
